backend/user_service/pyproject.toml
[tool.poetry]
name = "user_service"
version = "0.1.0"
description = ""
authors = ["Pavel Bocharov <skidyschggwp@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
djoser = "^2.2.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"
boto3 = "^1.34.145"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
